move testresult and failurereport to own files
