+++ src/suite/mod.rs
@@ -1,4 +1,5 @@
 pub mod extract;
+pub mod failure_report;
 pub mod headers;
 pub mod report;
 pub mod request;
@@ -6,5 +7,6 @@ pub mod response;
 pub mod setup;
 pub mod suite;
 pub mod test;
+pub mod test_result;
 
 pub use suite::Suite;
